Ability to cast by sharing url to homeassistant
The automation will capture the url from reading event mobile_app.share and update an input text. Then will trigger a script which will open the application and the content that is shared on google tv.

Currently supported applications are Netflix, PrimeVideo and YouTube.
To Do:
Hotstar (if I figure out the intent that will cast the content)